aigp-test/requirements.txt
gradio
torch
numpy
transformers
sentence-transformers
pandas
Pillow
accelerate